requirements.txt
discord.py==2.3.2
youtube-dl==2021.12.17
